REDES2/P3/requirements.txt
asgiref==3.8.1
Django==5.0.6
iniconfig==2.0.0
mock==5.1.0
packaging==24.1
paho-mqtt==2.1.0
pluggy==1.5.0
pytest==8.2.2
pytest-django==4.8.0
pytest-mock==3.14.0
sqlparse==0.5.0
